Scaffolded backend, adding new folders and blank files. Pulled utils and typed into relevant feature folders Also renamed existing files as [foldername].[relevant tag if any, like util].[ext]. this is the syntax now for all files, front and backend, through my codebase
