+++ .gitignore
@@ -14,6 +14,10 @@ dist
 dist-ssr
 *.local
 
+# don't distribute our DB layout lol
+build.md
+*.sql
+
 # Editor directories and files
 .vscode/*
 !.vscode/extensions.json
