--- server/build.txt
@@ -1,17 +1,0 @@
-What does my databse need?
-submitFiller - marathon {
-  lead manager Name
-  filler Name
-  ISV1
-  ISV2
-  BP
-  event
-}
-submitFiller - marathon {
-  lead manager Name
-  filler Name
-  ISV1
-  ISV2
-  BP
-  event
-}
