+++ server/routes/auth/auth.index.ts
