+++ src/components/Components.types.ts
@@ -1,0 +1,5 @@
+export const TeamFormFields = [
+  { id: 'isv1', stateKey: 'isv1', placeholder: 'ISV1' },
+  { id: 'isv2', stateKey: 'isv2', placeholder: 'ISV2' },
+  { id: 'bp', stateKey: 'bp', placeholder: 'BP' },
+];
