+++ src/components/Select.tsx
@@ -1,5 +1,5 @@
 import { Dispatch, SetStateAction } from 'react';
-import { handleStateChange } from '../features/create-run/utils';
+import { handleStateChange } from '../features/create-run/CreateRun.utils';
 
 interface SelectProps<T> {
   options: string[];
