+++ src/features/create-run/CreateRun.types.ts
@@ -1,0 +1,8 @@
+export type CreateRunFormState = {
+  selectedEvent: string;
+  runnerName: string;
+  isv1: number | undefined;
+  isv2: number | undefined;
+  bp: number | undefined;
+  runPassword: string;
+};
