+++ src/features/create-run/CreateRun.utils.ts
@@ -1,0 +1,10 @@
+export const runPasswordRegex = /^(?=.*[A-Z])(?=.*\d)(?=.*[\W_])[^\s]{12}$/;
+
+export const InitializeCreateRun = {
+  selectedEvent: '',
+  runnerName: '',
+  isv1: undefined,
+  isv2: undefined,
+  bp: undefined,
+  runPassword: '',
+};
