--- src/features/join-run/marathon/utils.ts
@@ -1,1 +1,0 @@
-//JULIA check with her to make sure these numbers are accurate
