+++ src/features/login/Login.types.ts
@@ -1,0 +1,9 @@
+export type LoginFormState = {
+  username: string;
+  password: string;
+};
+
+export const InitializeLogin = {
+  username: '',
+  password: '',
+};
