+++ src/features/register/Register.types.ts
@@ -1,0 +1,11 @@
+export type RegisterFormState = {
+  username: string;
+  password: string;
+  passwordConfirm: string;
+};
+
+export const InitializeRegister = {
+  username: '',
+  password: '',
+  passwordConfirm: '',
+};
