+++ src/features/run-data/components/TeamContainer.tsx
@@ -1,7 +1,11 @@
 import { SetStateAction, useState } from 'react';
 import TeamInputs from './TeamInputs';
 import { Teams, Filler } from '@/utils/types';
-import { compareFillersAndTeams, createTempTeams, Diff } from '../utils';
+import {
+  compareFillersAndTeams,
+  createTempTeams,
+  Diff,
+} from '../RunData.utils';
 import { DummyFillerData } from '@/utils/variables';
 import SubmitModal from './modals/SubmitModal';
 import CancelModal from './modals/CancelModal';
