+++ src/features/run-data/components/modals/SubmitModal.tsx
@@ -1,7 +1,7 @@
 import Modal from '@/components/Modal';
 import { Filler, Teams } from '@/utils/types';
 import { teamParentMap } from '@/utils/variables';
-import { Diff } from '../../utils';
+import { Diff } from '../../RunData.utils';
 import { SetStateAction } from 'react';
 
 interface SubmitModalProps {
