+++ src/utils/dummyData.db.ts
@@ -1,0 +1,143 @@
+export const DummyFillerData = [
+  {
+    name: 'Ki',
+    teams: {
+      healTeam: { isv1: 13, isv2: 48, bp: 7 },
+      fillTeam: { isv1: 47, isv2: 9, bp: 30 },
+      sb1: { isv1: 16, isv2: 43, bp: 16 },
+      sb2: { isv1: 16, isv2: 32, bp: 8 },
+    },
+    event: 'Moonlight Dash',
+    leadManager: 'Chris',
+  },
+  {
+    name: 'Zach',
+    teams: {
+      healTeam: { isv1: 17, isv2: 29, bp: 30 },
+      fillTeam: { isv1: 35, isv2: 30, bp: 31 },
+      sb1: { isv1: 6, isv2: 46, bp: 4 },
+      sb2: { isv1: 3, isv2: 7, bp: 40 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Casey',
+  },
+  {
+    name: 'Jess',
+    teams: {
+      healTeam: { isv1: 25, isv2: 38, bp: 49 },
+      fillTeam: { isv1: 24, isv2: 1, bp: 5 },
+      sb1: { isv1: 29, isv2: 26, bp: 12 },
+      sb2: { isv1: 21, isv2: 19, bp: 41 },
+    },
+    event: 'Starry Night',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Alexr',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Aletx',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Aleyx',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Aluex',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Aleix',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Aleox',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Alpex',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Ale[x',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+  {
+    name: 'Ale]x',
+    teams: {
+      healTeam: { isv1: 15, isv2: 21, bp: 6 },
+      fillTeam: { isv1: 18, isv2: 45, bp: 9 },
+      sb1: { isv1: 20, isv2: 10, bp: 35 },
+      sb2: { isv1: 10, isv2: 23, bp: 46 },
+    },
+    event: 'Footsteps',
+    leadManager: 'Jordan',
+  },
+];
+
+export const DummyManagerData = [
+  { username: 'Eli', id: 'dkhfdlashfasdfkaslfkh' },
+  { username: 'Levi', id: 'dkhfdlashfasdfkaslfkh' },
+  { username: 'Ali', id: 'dkhfdlashfasdfkaslfkh' },
+  { username: 'Ellie', id: 'dlsfgjd;fsogjidfsgsdfg' },
+  { username: 'Mikui', id: 'rewt89540uiorwtgjr;ew' },
+  { username: 'Leo', id: ';cvmklbclxkbm' },
+];
